flask-project/requirements.txt
Flask==2.2.3
py_cpuinfo==9.0.0
